[REFACTOR] Moves expression of interest test assert functions wider accessible
